Refactor packages to prince package
