--- src/test/java/Tasks/TaskListTest.java
@@ -1,4 +1,0 @@
-package tasks;
-
-public class TaskListTest {
-}
