+++ src/test/java/prince/tasks/TaskListTest.java
@@ -1,0 +1,4 @@
+package prince.tasks;
+
+public class TaskListTest {
+}
