+++ src/test/java/prince/util/ParserTest.java
@@ -1,0 +1,4 @@
+package prince.util;
+
+public class ParserTest {
+}
