+++ src/test/java/prince/util/UiTest.java
@@ -1,0 +1,4 @@
+package prince.util;
+
+public class UiTest {
+}
