--- src/test/java/util/ParserTest.java
@@ -1,4 +1,0 @@
-package util;
-
-public class ParserTest {
-}
