--- src/test/java/util/UiTest.java
@@ -1,4 +1,0 @@
-package util;
-
-public class UiTest {
-}
